% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_density_h2o.R
\name{calc_density_h2o}
\alias{calc_density_h2o}
\title{Density of water}
\usage{
calc_density_h2o(tc, patm)
}
\arguments{
\item{tc}{Air temperature (degrees Celsius)}

\item{patm}{Atmospheric pressure (Pa)}
}
\description{
Calculates the density of water at a given temperature and pressure. Adopted
from implementation by Davis et al. (2017).
}
\details{
Returns the density of water (kg m-3)
}
\references{
Chen et al. (1977)

Davis, T. W., Prentice, I. C., Stocker, B. D., Thomas, R. T., Whitley, R. J.,
Wang, H., Evans, B. J., Gallego-Sala, A. V., Sykes, M. T., & Cramer, W.
(2017). Simple process-led algorithms for simulating habitats (SPLASH v.1.0):
Robust indices of radiation, evapotranspiration and plant-available moisture.
Geoscientific Model Development, 10(2), 689–708.
https://doi.org/10.5194/gmd-10-689-2017
}
