% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_et.R
\name{convert_et}
\alias{convert_et}
\title{Water energy to mass conversion}
\usage{
convert_et(et_e, tc, patm, return_df = FALSE)
}
\arguments{
\item{et_e}{Latent heat flux in energy units (W m-2; representative for the
mean value over one time step)}

\item{tc}{Air temperature in degrees Celsius}

\item{patm}{Atmospheric pressure (Pa)}

\item{return_df}{A logical specifying whether to return a data frame with a
single column containing the water mass flux. Defaults to \code{FALSE}}
}
\description{
Convert the latent heat flux into evapotranspiration in (water) mass units.
Based on the SPLASH model (Davis et al. 2017 http://www.geosci-model-dev.net/10/689/2017/)
}
\details{
Returns the water mass flux in mm d-1 (equivalent to kg m-2)
}
