% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_snow.R
\name{simulate_snow}
\alias{simulate_snow}
\title{Simulate snow mass}
\usage{
simulate_snow(df, varnam_temp, varnam_prec, varnam_snow)
}
\arguments{
\item{df}{A data frame containing columns for air temperature (deg. C),
precipitation in liquid form (rain, mm d-1), and precipitation in solid form
(snow water equivalents, mm d-1). The column names of the respective
variables are provided by the other arguments. Data frame must contain at
least 365 days (i.e. a full year). It can, however, contain more e.g. 400 days,
i.e. the current year still ongoing.}

\item{varnam_temp}{A character string specifying the variable name for air
temperature.}

\item{varnam_prec}{A character string specifying the variable name for rain.}

\item{varnam_snow}{A character string specifying the variable name for snow.}
}
\description{
Simulate snow mass accumulation and melt based on Orth et al. (2013)
https://www.jstor.org/stable/24914341
}
\details{
Returns a data frame with two added columns: (1) \code{liquid_to_soil}
is the rain plus snow melt in mm d-1; (2) \code{snow_pool} is the snow mass
in water equivalents (mm) for each day.
}
