% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_enthalpy_vap.R
\name{calc_enthalpy_vap}
\alias{calc_enthalpy_vap}
\title{Enthalpy of vaporisation}
\usage{
calc_enthalpy_vap(tc)
}
\arguments{
\item{tc}{Air temperature in degrees Celsius}
}
\description{
Calculates the enthalpy of vaporisation for converting a mass (flux) into an
energy (flux). The enthalpy of vaporization refers to the amount of heat
required to change a substance from liquid to vapor at constant pressure. It
includes both the internal energy required to overcome intermolecular forces
within the liquid and the work done against the surrounding atmosphere.
Enthalpy of vaporization and latent heat of vaporization are often used
interchangeably, but there is a subtle difference between them. Latent heat
of vaporization, on the other hand, specifically refers to the heat required
to change a unit mass of a substance from liquid to vapor at its boiling
point at constant temperature. It represents the energy needed to overcome
intermolecular forces alone.
}
\details{
Returns the enthalpy of vaporisation (J/kg)
}
\references{
Henderson-Sellers (1984)

Davis, T. W., Prentice, I. C., Stocker, B. D., Thomas, R. T., Whitley, R. J.,
Wang, H., Evans, B. J., Gallego-Sala, A. V., Sykes, M. T., & Cramer, W.
(2017). Simple process-led algorithms for simulating habitats (SPLASH v.1.0):
Robust indices of radiation, evapotranspiration and plant-available moisture.
Geoscientific Model Development, 10(2), 689–708.
https://doi.org/10.5194/gmd-10-689-2017
}
