% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R
\name{pet}
\alias{pet}
\title{Priestley-Taylor potential evapotranspiration}
\usage{
pet(netrad, tc, patm, return_df = FALSE)
}
\arguments{
\item{netrad}{Net radiation (W m-2)}

\item{tc}{Air temperature in degrees Celsius}

\item{patm}{Atmospheric pressure (Pa)}

\item{return_df}{A logical specifying whether to return a data frame with a
single column containing potential evapotranspiration. Defaults to \code{FALSE}}
}
\description{
Calculates the potential evapotranspiration using the implementation by
Davis et al (2017), based on Priestley & Taylor (1972)
}
\details{
Potential evapotranspiration (mm s-1)
}
\references{
Davis, T. W., Prentice, I. C., Stocker, B. D., Thomas, R. T., Whitley, R. J.,
Wang, H., Evans, B. J., Gallego-Sala, A. V., Sykes, M. T., & Cramer, W.
(2017). Simple process-led algorithms for simulating habitats (SPLASH v.1.0):
Robust indices of radiation, evapotranspiration and plant-available moisture.
Geoscientific Model Development, 10(2), 689–708.
https://doi.org/10.5194/gmd-10-689-2017

Priestley, C. H. B., & Taylor, R. J. (1972). On the Assessment of Surface
Heat Flux and Evaporation Using Large-Scale Parameters. Monthly Weather
Review, 100(2), 81–92. https://doi.org/10.1175/1520-0493
}
